Also reopen the Get This Book! modal when retrying Stale Element exceptions
