+++ pages/webview/content.py
@@ -4,11 +4,13 @@
 
 import re
 
+from functools import wraps
+
 from selenium.webdriver.common.by import By
+from selenium.common.exceptions import StaleElementReferenceException
 
 from tests.utils import retry_stale_element_reference_exception
 
-from decorators.exception_handling import retry_stale_element_reference_exception
 from pages.webview.base import Page
 from regions.webview.base import Region
 from regions.webview.content_item import ContentItem
@@ -44,6 +46,27 @@ class Content(Page):
     @property
     def share(self):
         return self.content_header.share
+
+    # Retries StaleElementReferenceExceptions up to n-1 times (default n=3)
+    # Decorator with optional argument based on: https://stackoverflow.com/a/3931903
+    def retry_stale_element_reference_exception(method_or_max_tries):
+        def wrap(method):
+            @wraps(method)
+            def wrapper(*args, **kwargs):
+                for i in range(max_tries):
+                    try:
+                        return method(*args, **kwargs)
+                    except StaleElementReferenceException:
+                        if i >= max_tries - 1:
+                            raise
+            return wrapper
+
+        if callable(method_or_max_tries):
+            max_tries = 3
+            return wrap(method_or_max_tries)
+        else:
+            max_tries = method_or_max_tries
+            return wrap
 
     # The media-header div can be reloaded at seemingly random times
     # Any method that accesses an element inside this header
@@ -106,7 +129,7 @@ class Content(Page):
     @retry_stale_element_reference_exception
     def click_get_this_book_button(self):
         self.get_this_book_button.click()
-        return self.GetThisBook(self).wait_for_region_to_display()
+        return self.GetThisBook(self).wait_for_region_to_load()
 
     class ContentHeader(Region):
         _root_locator = (By.CSS_SELECTOR, '#content div.pinnable')
@@ -307,8 +330,9 @@ class Content(Page):
                             self.root.click()
                             return self.page.wait_for_url_to_change(current_url)
 
-    # This entire region can be overwritten at any time,
-    # so ALL methods that access the DOM must retry StaleElementReferenceExceptions
+    # This entire region can be overwritten (and the modal automatically closed) at any time,
+    # so any tests that use it must be ready to retry StaleElementReferenceExceptions
+    # including clicking the "Get This Book!" button again
     class GetThisBook(Region):
         _root_locator = (By.CSS_SELECTOR, 'div.popover div.popover-content div.book-popover')
         _pdf_link_locator = (
@@ -320,39 +344,46 @@ class Content(Page):
         _order_printed_book_link_locator = (
             By.CSS_SELECTOR, 'a.order[data-l10n-id="textbook-view-book-order-book"]')
 
+        # Retries StaleElementReferenceExceptions up to n-1 times (default n=3)
+        # Decorator with optional argument based on: https://stackoverflow.com/a/3931903
+        # This version also reopens the GetThisBook modal, since it closes when this happens
+        retry_stale_element_reference_exception_and_reopen_modal = (
+            retry_stale_element_reference_exception(
+                3, lambda arg, kwarg: ((arg[0].page.click_get_this_book_button(), arg[1:]), kwarg)))
+
         @property
-        @retry_stale_element_reference_exception
+        @retry_stale_element_reference_exception_and_reopen_modal
         def root(self):
             return super().root
 
         @property
-        @retry_stale_element_reference_exception
+        @retry_stale_element_reference_exception_and_reopen_modal
         def text(self):
             return super().text
 
         @property
-        @retry_stale_element_reference_exception
+        @retry_stale_element_reference_exception_and_reopen_modal
         def loaded(self):
             return super().loaded
 
         @property
-        @retry_stale_element_reference_exception
+        @retry_stale_element_reference_exception_and_reopen_modal
         def is_displayed(self):
             return super().is_displayed
 
-        @retry_stale_element_reference_exception
+        @retry_stale_element_reference_exception_and_reopen_modal
         def find_element(self, strategy, locator):
             return super().find_element(strategy, locator)
 
-        @retry_stale_element_reference_exception
+        @retry_stale_element_reference_exception_and_reopen_modal
         def find_elements(self, strategy, locator):
             return super().find_elements(strategy, locator)
 
-        @retry_stale_element_reference_exception
+        @retry_stale_element_reference_exception_and_reopen_modal
         def is_element_present(self, strategy, locator):
             return super().is_element_present(strategy, locator)
 
-        @retry_stale_element_reference_exception
+        @retry_stale_element_reference_exception_and_reopen_modal
         def is_element_displayed(self, strategy, locator):
             return super().is_element_displayed(strategy, locator)
